# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce Order & Promotion API",
        "description": "REST API for browsing products, managing a cart, applying promotional codes, checking out and managing orders.\n\n**Money** is returned as decimal strings (e.g. `\"149.99\"`) and accepted in dollars; it is stored as integer cents.\n\n**Errors** share one shape: `{ \"message\", \"code\", \"details\"? }`. Validation errors (`validation_failed`) add an `errors` object keyed by field.\n\n**Authentication**: call `POST /api/auth/login` (or register), then send `Authorization: Bearer <token>`.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "This server"
        }
    ],
    "paths": {
...
        "/api/cart/promotion": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Apply a promotional code",
                "description": "Codes are matched ignoring case and surrounding spaces. Eligibility is checked again at checkout.",
                "operationId": "applyPromotion",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "SUMMER20",
                                        "maxLength": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The cart with the discount applied.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Cart"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "422": {
                        "description": "The customer is not eligible: `coupon_invalid` (unknown, inactive or not started), `coupon_expired`, `coupon_usage_limit_reached`, `coupon_customer_limit_reached`, `coupon_minimum_not_met`; or `cart_empty`; or `validation_failed`.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Error"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ValidationError"
                                        }
                                    ]
                                },
                                "examples": {
                                    "coupon_invalid": {
                                        "summary": "coupon_invalid",
                                        "value": {
                                            "message": "The coupon code is invalid.",
                                            "code": "coupon_invalid"
                                        }
                                    },
                                    "coupon_expired": {
                                        "summary": "coupon_expired",
                                        "value": {
                                            "message": "This coupon has expired.",
                                            "code": "coupon_expired"
                                        }
                                    },
                                    "coupon_usage_limit_reached": {
                                        "summary": "coupon_usage_limit_reached",
                                        "value": {
                                            "message": "This coupon has reached its usage limit.",
                                            "code": "coupon_usage_limit_reached"
                                        }
                                    },
                                    "coupon_customer_limit_reached": {
                                        "summary": "coupon_customer_limit_reached",
                                        "value": {
                                            "message": "You have already used this coupon the maximum number of times.",
                                            "code": "coupon_customer_limit_reached"
                                        }
                                    },
                                    "coupon_minimum_not_met": {
                                        "summary": "coupon_minimum_not_met",
                                        "value": {
                                            "message": "This coupon requires a minimum cart subtotal of 100.00.",
                                            "code": "coupon_minimum_not_met",
                                            "details": {
                                                "minimum": "100.00",
                                                "subtotal": "99.99"
                                            }
                                        }
                                    },
                                    "cart_empty": {
                                        "summary": "cart_empty",
                                        "value": {
                                            "message": "Your cart is empty.",
                                            "code": "cart_empty"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Remove the promotional code",
                "operationId": "removePromotion",
                "responses": {
                    "200": {
                        "description": "The cart without a promotion.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Cart"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/checkout": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Place an order from the cart",
                "description": "Takes no body: prices and totals are recalculated on the server from the locked product rows, and anything sent by the client is ignored. Runs in one database transaction with row locks, so concurrent checkouts can never oversell stock or a coupon. Send an Idempotency-Key header to make retries safe.",
                "operationId": "checkout",
                "parameters": [
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "description": "Optional, up to 100 characters (letters, numbers, - _ : .). Retrying with the same key returns the original order with status 200.",
                        "schema": {
                            "type": "string",
                            "example": "checkout-6f1c2a",
                            "maxLength": 100
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Order placed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Replay of an earlier request with the same Idempotency-Key; nothing new was placed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "422": {
                        "description": "Nothing was changed: `cart_empty`, `insufficient_stock` (lists every short item), a coupon error (`coupon_*`), or an invalid Idempotency-Key (`validation_failed`).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Error"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ValidationError"
                                        }
                                    ]
                                },
                                "examples": {
                                    "cart_empty": {
                                        "summary": "cart_empty",
                                        "value": {
                                            "message": "Your cart is empty.",
                                            "code": "cart_empty"
                                        }
                                    },
                                    "insufficient_stock": {
                                        "summary": "insufficient_stock",
                                        "value": {
                                            "message": "Only 3 units of 'Desk Lamp' are available.",
                                            "code": "insufficient_stock",
                                            "details": {
                                                "items": [
                                                    {
                                                        "product_id": 1,
                                                        "name": "Desk Lamp",
                                                        "requested": 4,
                                                        "available": 3
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "coupon_expired": {
                                        "summary": "coupon_expired",
                                        "value": {
                                            "message": "This coupon has expired.",
                                            "code": "coupon_expired"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "required": [
                    "message",
                    "code"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Only 3 units of 'Desk Lamp' are available."
                    },
                    "code": {
                        "description": "Machine-readable error code.",
                        "type": "string",
                        "example": "insufficient_stock"
                    },
                    "details": {
                        "description": "Extra context for some errors.",
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": true
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "required": [
                    "message",
                    "code",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The quantity field must be at least 1."
                    },
                    "code": {
                        "type": "string",
                        "example": "validation_failed"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "quantity": [
                                "The quantity field must be at least 1."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "CartItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 7
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Walnut Desk Lamp"
                    },
                    "sku": {
                        "type": "string",
                        "example": "LAMP-001"
                    },
                    "unit_price": {
                        "description": "Current product price.",
                        "type": "string",
                        "example": "49.99"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "line_total": {
                        "type": "string",
                        "example": "99.98"
                    },
                    "available_quantity": {
                        "description": "Units that can be bought now (0 if the product was deactivated).",
                        "type": "integer",
                        "example": 5
                    },
                    "is_available": {
                        "description": "False when stock dropped below the quantity or the product was deactivated.",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "Cart": {
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CartItem"
                        }
                    },
                    "item_count": {
                        "type": "integer",
                        "example": 3
                    },
                    "subtotal": {
                        "type": "string",
                        "example": "154.00"
                    },
                    "promotion": {
                        "description": "The applied code, re-evaluated on every read. An invalid code is reported and not discounted.",
                        "properties": {
                            "code": {
                                "type": "string",
                                "example": "SUMMER20"
                            },
                            "type": {
                                "type": "string",
                                "enum": [
                                    "percentage",
                                    "fixed"
                                ]
                            },
                            "is_valid": {
                                "type": "boolean"
                            },
                            "error": {
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "coupon_minimum_not_met"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "discount": {
                        "type": "string",
                        "example": "30.80"
                    },
                    "total": {
                        "type": "string",
                        "example": "123.20"
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "description": "Snapshot taken at purchase time; later product changes do not affect it.",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "product_id": {
                        "description": "Null if the product was later deleted.",
                        "type": "integer",
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Walnut Desk Lamp"
                    },
                    "sku": {
                        "type": "string",
                        "example": "LAMP-001"
                    },
                    "unit_price": {
                        "type": "string",
                        "example": "49.99"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "line_total": {
                        "type": "string",
                        "example": "99.98"
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 42
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "processing",
                            "shipped",
                            "delivered",
                            "cancelled"
                        ]
                    },
                    "subtotal": {
                        "type": "string",
                        "example": "154.00"
                    },
                    "discount": {
                        "type": "string",
                        "example": "30.80"
                    },
                    "total": {
                        "type": "string",
                        "example": "123.20"
                    },
                    "promotion_code": {
                        "type": "string",
                        "example": "SUMMER20",
                        "nullable": true
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    },
                    "cancelled_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "Unauthenticated": {
                "description": "Missing or invalid bearer token (`unauthenticated`).",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "example": {
                            "message": "Unauthenticated.",
                            "code": "unauthenticated"
                        }
                    }
                }
            },
...
            "TooManyRequests": {
                "description": "Rate limit exceeded (`too_many_requests`); see the `Retry-After` header.",
                "headers": {
                    "Retry-After": {
                        "description": "Seconds until another request is allowed.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                },
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "example": {
                            "message": "Too Many Attempts.",
                            "code": "too_many_requests"
                        }
                    }
                }
            }
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum personal access token from /api/auth/login or /api/auth/register.",
                "bearerFormat": "Sanctum token",
                "scheme": "bearer"
            }
        }
    },
...
}